pyproject.toml
[project]
name = "yolo-demo1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ultralytics>=8.0.0",
    "numpy>=1.24.0",
    "torch>=2.7.1+cu128",
    "torchvision>=0.22.1+cu128",
    "torchaudio>=2.7.1+cu128",
    "Pillow>=9.5.0",
    "PyYAML>=6.0",
    "psutil>=5.9.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]


# [[tool.uv.index]]
# name = "pytorch"
# url = "https://download.pytorch.org/whl/cu128"
# explicit = true